+++ src/db/schema/pdf/highlights.ts
@@ -1,4 +1,4 @@
-import { index, integer, jsonb, pgTable, text } from "drizzle-orm/pg-core";
+import { index, jsonb, pgTable, smallint, text } from "drizzle-orm/pg-core";
 import { pdfsTable } from "@/db/schema/pdf/pdfs";
 import type { Rect } from "@/lib/types";
 
@@ -9,7 +9,7 @@ export const highlightsTable = pgTable(
     rects: jsonb("rects").$type<Rect[]>().notNull(),
     color: text("color").notNull(),
     text: text("text").notNull(),
-    pageNumber: integer("page_number").notNull(),
+    pageNumber: smallint("page_number").notNull(),
     pdfId: text("pdf_id")
       .references(() => pdfsTable.id, { onDelete: "cascade" })
       .notNull(),
